pyproject.toml
[tool.poetry]
name = "finops-report-automation"
version = "0.1.0"
description = "Simple tool aimed to create automated cloud usage optimization reports"
authors = [
  "Josha Voller <j.voller@reply.de>",
  "Cristobal Abarca <c.abarcaquiroga@reply.de>"
  ]

[tool.poetry.dependencies]
python = "^3.8.10"
PyYAML = "5.4.1"
openpyxl = "3.0.9"
requests = "^2.27.1"
pytest = "^7.1.1"
flake8 = "^4.0.1"
pandas = "^1.4.2"
Jinja2 = "^3.1.2"
boto3 = "^1.24.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
